Fixed full text parsing sporadically showing error message
New article view model. Fixes NOt HMTL error

New feed view model

Cleaned up old viewmodel stuff
