+++ app/src/main/java/com/nononsenseapps/feeder/di/ArchModelModule.kt
@@ -16,7 +16,8 @@ import com.nononsenseapps.feeder.ui.NavigationDeepLinkViewModel
 import com.nononsenseapps.feeder.ui.OpenLinkInDefaultActivityViewModel
 import com.nononsenseapps.feeder.ui.compose.editfeed.CreateFeedScreenViewModel
 import com.nononsenseapps.feeder.ui.compose.editfeed.EditFeedScreenViewModel
-import com.nononsenseapps.feeder.ui.compose.feedarticle.FeedArticleViewModel
+import com.nononsenseapps.feeder.ui.compose.feedarticle.ArticleViewModel
+import com.nononsenseapps.feeder.ui.compose.feedarticle.FeedViewModel
 import com.nononsenseapps.feeder.ui.compose.searchfeed.SearchFeedViewModel
 import com.nononsenseapps.feeder.ui.compose.settings.SettingsViewModel
 import org.kodein.di.DI
@@ -41,6 +42,7 @@ val archModelModule =
         bindWithComposableViewModelScope<EditFeedScreenViewModel>()
         bindWithComposableViewModelScope<CreateFeedScreenViewModel>()
         bindWithComposableViewModelScope<SearchFeedViewModel>()
-        bindWithComposableViewModelScope<FeedArticleViewModel>()
+        bindWithComposableViewModelScope<ArticleViewModel>()
+        bindWithComposableViewModelScope<FeedViewModel>()
         bindWithComposableViewModelScope<NavigationDeepLinkViewModel>()
     }
